+++ test/unit/fuzz-tools.test.js
@@ -363,6 +363,7 @@ import * as risserSignV383 from '../../lib/risser-sign-v383.js';
 import * as spetzlerPonceV384 from '../../lib/spetzler-ponce-v384.js';
 import * as schwabEnglandV385 from '../../lib/schwab-england-v385.js';
 import * as piraniClubfootV386 from '../../lib/pirani-clubfoot-v386.js';
+import * as dimeglioClubfootV387 from '../../lib/dimeglio-clubfoot-v387.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -671,6 +672,7 @@ const MODULES = {
   'spetzler-ponce-v384.js': spetzlerPonceV384,
   'schwab-england-v385.js': schwabEnglandV385,
   'pirani-clubfoot-v386.js': piraniClubfootV386,
+  'dimeglio-clubfoot-v387.js': dimeglioClubfootV387,
   'tb-testing.js': tbTesting,
 };